lending-product/src/main/java/com/fusionx/lending/product/resources/EligibilityOtherUpdateResource.java
package com.fusionx.lending.product.resources;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.Pattern;
import javax.validation.constraints.Size;

import com.fasterxml.jackson.databind.annotation.JsonSerialize;


/**
 * EligibilityOtherUpdateResource
 * 
 *******************************************************************************************************
 *  ###   Date         Story Point   		Task No    Author       Description
 *------------------------------------------------------------------------------------------------------
 *    1   27-07-2021    FXL_July_2021_2  	FXL-54		Piyumi      Created
 *    
 *******************************************************************************************************
 */

@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
public class EligibilityOtherUpdateResource {
	
	@NotBlank(message = "{common.not-null}")
	@Pattern(regexp = "^$|[0-9]+", message = "{common-numeric.pattern}")
	private String id;
	
	@Pattern(regexp = "^$|[0-9]+", message = "{common-numeric.pattern}")
	private String eligibilityPendingId;
	
	@NotBlank(message = "{common.not-null}")
	@Pattern(regexp = "^$|[0-9]+", message = "{common-numeric.pattern}")
	private String otherEligibilityTypeId;
	
	@NotBlank(message = "{common.not-null}")
	@Size(max = 70, message = "{common-name.size}")
	private String otherEligibilityTypeName;
	
	@NotBlank(message = "{common.not-null}")
	@Pattern(regexp = "ACTIVE|INACTIVE", message = "{common-status.pattern}")
	private String status;
	
	@NotBlank(message = "{common.not-null}")
	@Pattern(regexp = "^$|[0-9]+", message = "{common-numeric.pattern}")
	private String version;
	
	public String getId() {
		return id;
	}

	public void setId(String id) {
		this.id = id;
	}
	
	public String getEligibilityPendingId() {
		return eligibilityPendingId;
	}

	public void setEligibilityPendingId(String eligibilityPendingId) {
		this.eligibilityPendingId = eligibilityPendingId;
	}
	
	public String getOtherEligibilityTypeId() {
		return otherEligibilityTypeId;
	}

	public void setOtherEligibilityTypeId(String otherEligibilityTypeId) {
		this.otherEligibilityTypeId = otherEligibilityTypeId;
	}
	
	public String getOtherEligibilityTypeName() {
		return otherEligibilityTypeName;
	}

	public void setOtherEligibilityTypeName(String otherEligibilityTypeName) {
		this.otherEligibilityTypeName = otherEligibilityTypeName;
	}

	public String getStatus() {
		return status;
	}

	public void setStatus(String status) {
		this.status = status;
	}
	
	public String getVersion() {
		return version;
	}

	public void setVersion(String version) {
		this.version = version;
	}

}
